Updated README content to accurately reflect reworked module
Pre-changes, this README reflected the current apache module. However,
the apache module is being reworked.

This is the updated README to reflect the workings of the new module.
It is only a first draft and will have more updates as questions are
answered and development continues. (This version still contains
self-notes. )