Updated content to conform to README best practices template
This README has gone through several drafts. Initially, it was the
module author-determined description of and instructions for use of the
module. I rewrote that README to reflect the Apache module as it was on
GitHub. After a discussion with @hunner , I reworked the README to
reflect his rewriting of the module.

This reworking added sections to reflect the new structuring of the
module, fleshed out content to explain the new features, added a ton of
parameters and examples to guide usage.

This commit is still a draft and is up for review.