Custom logging for apache::vhost
This patch adds the ability to change how vhost logging works. This patch
adds new parameters to 'apache::vhost' define type. This allows users to
specify a log format or nickname for the 'CustomLog' directive. This also
allows users to specify a pipe as the destination for the 'ErrorLog' and
'CustomLog' directives.