Move mpm_worker 'ServerLimit' line in ERB
It turns out that the order of MPM directives is important in certain
cases. Without this commit, the 'ServerLimit' directive was being
processed too late to be used in calculations for 'ThreadsPerChild'.

This commit moves the 'ServerLimit' directive to the top of the
directives in the ERB. This was based on Apache documentation and an
example config of theirs.