remove servername_real
previously, if you did this:
apache::vhost { 'foo.example.com':
    docroot => '/var/www/foo',
}

puppet would error, saying you need to specify servername and/or ip
and/or port.

but why?

just default to $name, as $servername_real was already sorta doing, but
it wasn't being fully utilized.

this commit gets rid of $servername_real and just makes $servername
default to $name if it's not specified, which is what I would expect it
to do anyways!