Support FallbackResource (httpd >= 2.2.16)
Starting Apache httpd 2.2.16, FallbackResource is a much simpler way of
configuring a very common pattern usually implemented with RewriteRules.

Sending all requests to one router that then dispaches all requests can
now be done with a single line:

    FallbackResource /index.php

We now support this directive with the apache::vhost parameter
fallbackresource, which can be used like this:

    apache::vhost { 'wordpress.example.com':
      documentroot     => '/var/www/wp',
      fallbackresource => '/index.php',
    }

fallbackresource is also a valid prameter in directories, where at
different levels it either can be overwritten, or `disabled`.