Resolves MODULES 1369 RedirectMatch rules do not work in Apache module
The existing template creates a line of the format
RedirectMatch permanent /(test)
This can not work with apache as the URL is a mandatory part of the redirect. To resolve this the patch creates a new variable redirectmatch_dest which follows the same flow as the redirectmatch_status and redirectmatch_regexp variables.
All three variables must be supplied and will generate a line of the format
RedirectMatch permanent /(test) http://www.example.com