Fix Shib setting rules.
ShibRequireSetting is not a real thing. I think it was a mixup between ShibRequireSession
and ShibRequestSetting. I've replaced it with separate options for both. The latter is
an array as you can set a variety of diffrent settings, while the latter is a simple
On or Off value. Each item in the settings array should be the name and value of the setting,
e.g. ['requireSession false', 'applicationId myresource']