MODULES-825:  apache 2.4  mod_itk needs mod_prefork
in apache 2.4 mod_prefork can be installed along with mod_itk

On ubuntu 14.04 mod_itk library name is "mpm_itk.so" instead of "mod_mpm_itk.so".
Added an if statement that creates a symlink in case of a debian distro. this preserve old manifest behaviour in library loading while fixing ubuntu bug.