Use string instead, not regex, for file test
The beaker-rspec contains matcher does not understand "+" in the
context of checking whether a string is contained in a file resource
(though it does seem to understand "\s" just fine), which was causing
acceptance test failures. Since we know the exact string to expect we
just remove the regular expression bits.