Specify owning permissions for logroot directory
The main goal of this patch is to prevent inheritance of owning
permissions for File resource. For example, if we define somewhere in
high level manifest 'owner' and 'group' values, it will be passed to
apache module and wrong owning permissions will be set. It's critical
for 'logroot' folder as we have '0750' permissions for this folder in
apache package by default.