Allow false booleans to set parameters to 'Off'
Without this change, explicitely setting these parameters to 'false'
does the same as not setting the parameters at all.

Since both parameters default to 'Off', this has no direct consequence,
but since we will introduce boolean that defaults to 'On', always output
the parameters with their value when they are set.