Enable the apache_version fact on FreeBSD
The fact was confined in 783729a5f2e348a4728f3ce29f0b58c7e0a425dd but
running the test suite on FreeBSD will fail because the fact is confined
to Linux.

Instead of simulating Linux when testing on FreeBSD, enable the fact on
FreeBSD so that the apache version is available in Puppet manifests on
this platform.