Use operatingsystemmajrelease fact
This fact was introduced in Facter 1.7 which by now can be assumed. By
now it's already deprecated in favor of $facts['os']['release']['major']
but the rest of the module isn't converted so this keeps it consistent.

CentOS 8 stream doesn't have a minor version and the previous comparison
code failed.

There are a lot more odd cases with facts, but I don't want to touch too
much code.