Add support for setting UserDir in Virtual Hosts
We can only set this setting at the VirtualHost level by relying on the
`custom_fragment` parameter, which does not scale when we want to enable
UserDir for a single VHost because we have to manually disable it on all
other VHosts.

With this change, one can ensure UserDir is active only on specific
VHost with something like:

```
Apache::Vhost {
  userdir => 'disabled',
}

apache::vhost { 'example.com':
  # [...]
  userdir => 'enabled',
}
```

A specific `userdir` parameter avoid clashes that would prevent this
technique to be effective with `custom_fragment`.