Automatically include modules used in vhost directories
The data passed to the directories parameter of apache::vhost often
implies a module should be present. This change inspects the passed data
and automatically includes modules if needed. This makes it easier to
use via Hiera while disabling all default mods.

Because of this change one acceptance test becomes obsolete: it no
longer fails and just works.