Group assertions in vhost spec
This uses chaining to remove a lot of separate expectations. The big
benefit is that it's a lot quicker to run and if one thing fails (like
the whole catalog fails to compile) there is just a single error for the
fragment rather than one for each test. It's also a lot shorter.