Allow custom_config to have a string priority again
In f41251e the type was narrowed to no
longer allow strings, but this can cause problems.

Sorting is alphabetical and you need to format it for the correct
sorting. So to make sure 2 loads before 10 you need to format it as 02.
While the vhost can do some printf style magic to change 2 to 02, it
must then also know what the highest number is. Otherwise 100 is sorted
before 20. By allowing strings, you allow the caller to fix this.

A type alias is introduced to reduce duplication and make it easier to
track.

Fixes: f41251e