Simplify the logic in _require.erb
If we take the expression:

    !(_item['require'] && _item['require'] != '') && _item['require'] !~ /unmanaged/i && !(_item['auth_require'])

This is hard to read, but equal to:

    (!_item['require'] || _item['require'] == '') && _item['require'] !~ /unmanaged/i && !(_item['auth_require'])

This makes it easier to see that item must be `nil` or `''`. Those will
never match `/unmanaged/i` so it can be simplified to:

    (!_item['require'] || _item['require'] == '') && !(_item['auth_require'])

Just before it is a block that already checks the last section, so it
can be reduced to an elsif on that block.